Refactor file download methods and update dependencies
Refactored the `HuggingFaceClient` class by replacing the `DownloadFileAsync` method with `DownloadFileWithResultAsync`, which returns a `FileDownloadResult` object. Added a new `GetDownloadUrl` method to generate download URLs. Updated the `IHuggingFaceClient` interface to reflect these changes.

Updated the `FileDownloadManager` and `RepositoryDownloadManager` classes to use the new `DownloadFileWithResultAsync` method and handle the `FileDownloadResult` object. Introduced the `FileDownloadResult` class to encapsulate download results.

Bumped the version of `Microsoft.Extensions.Logging.Abstractions` in the project file from `9.0.1` to `9.0.2`.